Initial. Broken currently

+++ Fusion-Core/src/graphics/buffers/indexbuffer.h
@@ -12,7 +12,7 @@ namespace fusion { namespace core { namespace graphics {
             GLuint m_Count;
 
         public:
-            IndexBuffer(GLushort* data, GLsizei count);
+            IndexBuffer(GLuint* data, GLsizei count);
 
             ~IndexBuffer();
