+++ Fusion-Core/src/graphics/static_sprite.cpp
@@ -25,7 +25,7 @@ namespace fusion { namespace core { namespace graphics {
 		m_VertexArray->addBuffer(new Buffer(vertices, 4 * 3, 3), 0);
 		m_VertexArray->addBuffer(new Buffer(colors, 4 * 4, 4), 1);
 		
-		GLushort indices[] = { 0, 1, 2, 2, 3, 0};
+		GLuint indices[] = { 0, 1, 2, 2, 3, 0};
 		m_IndexBuffer = new IndexBuffer(indices, 6);
 	}
